Crate sorted_array

This crate provides SortedSlice and SortedVec, which are sorted variants of [T] and Vec<T>.

These types have a restricted set of operations to guarantee that they are always sorted. For instance, it is not possible to get mutable references to the underlying slices or items within, as that could allow altering the array so that the items are no longer in order.

The types also have some additional functions that rely on the fact that the array is sorted to be more efficient that the standard slice functions.

Modules

prelude

Structs

OrdComparator

SortedSlice

SortedVec	feature="alloc"

Traits

Comparator

WeakBorrow

Functions

sort_boxed_slice	feature="alloc"

sort_mut_slice	feature="alloc"

sort_mut_slice_unstable

sort_vec